requirements.txt
fastapi
uvicorn
pandas
pillow
matplotlib